pyproject.toml
[project]
name = "teamfinder"
version = "0.1.0"
description = "Сервис поиска напарников для видеоигр"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2", # Unit-тесты
    "redis>=7.2.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[tool.setuptools.packages.find]
where = ["src"]
